There was an integer overflow in the way elapsed_time_ms.  Because the SGTimeStamp "-" operator returns it's result in usec's, there is an upper bound of 37.8 minutes on the maximum difference this operator can reliably report before the math overflows.  I change the global "int elapsed_time_ms" variable to a "double sim_time_ms" and restructured how the value is calculated. ---- The practical result of the overflow bug is that a large negative dt was passed to the event manager and the end effect was that no events would be run until their counters caught up ... in another 37.8 minutes or so.

+++ src/Main/logger.hxx
@@ -57,7 +57,7 @@ private:
     vector<SGPropertyNode *> nodes;
     ostream * output;
     long interval_ms;
-    long last_time_ms;
+    double last_time_ms;
     char delimiter;
   };
